refactor panel; start laying out upload buttons
